DataBase start

+++ app/src/main/java/by/hometrainng/databasehw4/model/ListElement.kt
@@ -1,7 +1,7 @@
 package by.hometrainng.databasehw4.model
 
 data class ListElement(
-    val id: Int = 1,
+    val id: Long = 1,
     val firstName: String = "Ivan",
     val lastName: String = "Ivanov",
 )
